auth added

+++ api/views.py
@@ -1,3 +1,4 @@
+# ---------------------------------------------------------------------------
 # from django.shortcuts import render
 # from .models import Book
 # from django.http import JsonResponse
@@ -62,8 +63,9 @@
 #             "delete": True,
 #         }, status=status.HTTP_404_NOT_FOUND)
 
+# ------------------------------------------------------------------------------------------------
+from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.views import APIView
-from rest_framework.decorators import api_view
 from rest_framework import status
 from rest_framework.response import Response
 from .models import Book
@@ -71,48 +73,71 @@ from .serializer import BookSerializer
 
 
 class Books(APIView):
-    def get(self, request):
-        try:
-            books = Book.objects.all()  # Complex data
-        except:
-            return Response({
-                "Erorr": "No book available",
-            }, status=status.HTTP_404_NOT_FOUND)
 
-        serializer = BookSerializer(books, many=True)
-        return Response(serializer.data)
+    def get(self, request):
+        token = request.COOKIES.get('jwt')
+        if token:
+            try:
+                books = Book.objects.all()  # Complex data
+            except:
+                return Response({
+                    "Erorr": "No book available",
+                }, status=status.HTTP_404_NOT_FOUND)
+
+            serializer = BookSerializer(books, many=True)
+            return Response(serializer.data)
+        else:
+            raise AuthenticationFailed('Unauthenticated!')
 
     def post(self, request):
-        book = BookSerializer(data=request.data)
-        if book.is_valid():
-            book.save()
-            return Response(book.data)
+        token = request.COOKIES.get('jwt')
+        if token:
+            book = BookSerializer(data=request.data)
+            if book.is_valid():
+                book.save()
+                return Response(book.data)
+            else:
+                return Response(book.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(book.errors, status=status.HTTP_400_BAD_REQUEST)
+            raise AuthenticationFailed('Unauthenticated!')
 
 
 class Update(APIView):
-    def get_book_by_pk(self,pk):
+
+    def get_book_by_pk(self, pk):
         book = Book.objects.get(pk=pk)
         return book
 
     def put(self, request, pk):
-        book = self.get_book_by_pk(pk)
-        s = BookSerializer(book, data=request.data)
-        if s.is_valid():
-            s.save()
-            return Response(s.data)
+        token = request.COOKIES.get('jwt')
+        if token:
+            book = self.get_book_by_pk(pk)
+            s = BookSerializer(book, data=request.data)
+            if s.is_valid():
+                s.save()
+                return Response(s.data)
+            else:
+                return Response(s.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(s.errors, status=status.HTTP_400_BAD_REQUEST)
+            raise AuthenticationFailed('Unauthenticated!')
 
     def get(self, request, pk):
-        book = book = self.get_book_by_pk(pk)
-        s = BookSerializer(book)
-        return Response(s.data)
+        token = request.COOKIES.get('jwt')
+        if token:
+            book = book = self.get_book_by_pk(pk)
+            s = BookSerializer(book)
+            return Response(s.data)
+        else:
+            raise AuthenticationFailed('Unauthenticated!')
 
     def delete(self, request, pk):
-        book = book = self.get_book_by_pk(pk)
-        book.delete()
-        return Response({
-            "delete": True,
-        }, status=status.HTTP_404_NOT_FOUND)
+        token = request.COOKIES.get('jwt')
+        if token:
+            book = book = self.get_book_by_pk(pk)
+            book.delete()
+            return Response({
+                "delete": True,
+            }, status=status.HTTP_404_NOT_FOUND)
+        else:
+            raise AuthenticationFailed('Unauthenticated!')
+# -----------------------------------------------------------------------------------------------
